1.71.9 (5.6.4f1)
Build-Id: 022fdeae-a616-499d-881d-0d429fe7e5e2
Build-Sha256: 35cc44f39b7cd83a76432dd6e519a7255b8b5cfc13813ec2a31dfea4a2d9936e

+++ Assembly-CSharp/CubeModelingTransitionTable.cs
@@ -12,7 +12,7 @@ internal class CubeModelingTransitionTable : StateTransitionTable
 		foreach (KeyValuePair<object, IState> item in table)
 		{
 			CubeModelTool cubeModelTool = (CubeModelTool)item.Value;
-			cubeModelTool.SetStateType((CubeModelingEvent)(int)item.Key);
+			cubeModelTool.SetStateType((CubeModelingEvent)item.Key);
 		}
 	}
 }
