+++ Assembly-CSharp/MVAvatar.cs
@@ -155,7 +155,7 @@ public class MVAvatar : MVGroup
 
 	protected virtual void AvatarStateChangedHandler(object a)
 	{
-		AvatarModeTypes avatarModeTypes = (AvatarModeTypes)(int)a;
+		AvatarModeTypes avatarModeTypes = (AvatarModeTypes)a;
 		if ((avatarModeTypes & AvatarModeTypes.Hidden) > AvatarModeTypes.None)
 		{
 			avatar.Collider.enabled = false;
