+++ Assembly-CSharp/NotificationTeamRequirementPanel.cs
@@ -23,7 +23,7 @@ public class NotificationTeamRequirementPanel : NotificationRequirementPanel
 	{
 		checkmark.sprite = checkmarkSprite;
 		textField.text = team.ToString();
-		switch ((MVTeam)(int)team)
+		switch ((MVTeam)team)
 		{
 		case MVTeam.Red:
 			requirementImage.sprite = redNotificationIcon;
